requirements.txt
# Django Core
Django>=5.0,<5.1
djangorestframework>=3.14.0

# Database
psycopg2-binary>=2.9.9  # PostgreSQL (للإنتاج)

# Caching & Background Tasks
redis>=5.0.0
celery>=5.3.0
django-redis>=5.4.0

# Forms & UI
django-crispy-forms>=2.1
crispy-bootstrap5>=2.0.0
django-filter>=23.5

# Import/Export
django-import-export>=3.3.0
openpyxl>=3.1.2
xlsxwriter>=3.1.9

# Images
Pillow>=10.1.0

# Authentication & Permissions
django-allauth>=0.57.0
django-guardian>=2.4.0

# Notifications
django-notifications-hq>=1.8.3

# Internationalization
django-modeltranslation>=0.18.11

# PDF Generation
WeasyPrint>=60.1
reportlab>=4.0.7

# Environment Variables
python-decouple>=3.8

# Cloud Storage (Optional)
django-storages>=1.14.2
boto3>=1.34.0

# Development Tools
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3

# Database URL
dj-database-url>=2.1.0

# CORS Headers
django-cors-headers>=4.3.1

# Timezone
pytz>=2023.3

# Utilities
python-dateutil>=2.8.2

